+++ e2e-tests/src/lib.rs
@@ -2,6 +2,9 @@
 mod azero;
 
 #[cfg(test)]
+mod balance;
+
+#[cfg(test)]
 mod client;
 
 #[cfg(test)]
