+++ e2e-tests/src/test/azero_to_eth.rs
@@ -9,9 +9,8 @@ use crate::{client::Client, config::setup_test, wait::wait_for_balance_change};
 /// This is easily done by running the test for the other direction first.
 /// Approves the `most` contract to use the wETH funds.
 /// Burns the required funds in the wETH contract on Aleph Zero.
-/// Transfers `transfer_amount` of burned wETH over the bridge, unwrapping the transfer to a specified Ethereum account.
+/// Transfers `transfer_amount` of burned wETH over the bridge.
 /// Waits for the transfer to complete.
-/// Verifies that the correct amount of ETH is present on the Ethereum chain.
 /// It relies on all the relevant contracts being deployed on both ends and the (wETH_ETH:wETH_AZERO) pair having been added to `most`.
 #[tokio::test]
 pub async fn weth_to_weth() -> Result<()> {
@@ -21,9 +20,9 @@ pub async fn weth_to_weth() -> Result<()> {
     let client = Client::new(test_context);
     let initial_balance = client.balance().await?;
 
-    info!("{:?}", initial_balance);
+    info!("Initial balance: {:?}", initial_balance);
 
-    info!("Approve the `most` contract to use the wETH funds on the Azero chain");
+    info!("Approve the `most` contract to use the wETH funds on the Aleph chain");
     client.approve_weth_azero(transfer_amount).await?;
 
     info!("Request the transfer of wETH to the Ethereum chain");
@@ -44,37 +43,14 @@ pub async fn weth_to_weth() -> Result<()> {
     .await
 }
 
-#[tokio::test]
-pub async fn weth_to_eth() -> Result<()> {
-    let config = setup_test();
-    let test_context = config.create_test_context().await?;
-    let transfer_amount = utils::parse_ether(config.test_args.transfer_amount.clone())?.as_u128();
-    let client = Client::new(test_context);
-    let initial_balance = client.balance().await?;
-
-    info!("{:?}", initial_balance);
-
-    info!("Approve the `most` contract to use the wETH funds on the Azero chain");
-    client.approve_weth_azero(transfer_amount).await?;
-
-    info!("Request the transfer of wETH to the Ethereum chain");
-    client.request_eth_transfer_azero(transfer_amount).await?;
-
-    info!("Wait for balance change");
-    let target_balance = initial_balance.bridge_eth_azero_to_eth(transfer_amount)?;
-    info!("Target balance: {:?}", target_balance);
-
-    let get_current_balance = || async { client.balance().await };
-    wait_for_balance_change(
-        get_current_balance,
-        target_balance,
-        Some(0.into()),
-        None,
-        config.test_args.wait_max_minutes,
-    )
-    .await
-}
-
+/// One-way `Aleph Zero` -> `Ethereum` transfer through `most`.
+/// Requires a prior transaction in the other direction to have completed.
+/// This is easily done by running the test for the other direction first.
+/// Approves the `most` contract to use the USDT funds.
+/// Burns the required funds in the USDT contract on Aleph Zero.
+/// Transfers `transfer_amount` of burned USDT over the bridge.
+/// Waits for the transfer to complete.
+/// It relies on all the relevant contracts being deployed on both ends and the (USDT_ETH:USDT_AZERO) pair having been added to `most`.
 #[tokio::test]
 pub async fn usdt_to_usdt() -> Result<()> {
     let config = setup_test();
@@ -83,9 +59,9 @@ pub async fn usdt_to_usdt() -> Result<()> {
     let client = Client::new(test_context);
     let initial_balance = client.balance().await?;
 
-    info!("{:?}", initial_balance);
+    info!("Initial balance: {:?}", initial_balance);
 
-    info!("Approve the `most` contract to use the USDT funds on the Azero chain");
+    info!("Approve the `most` contract to use the USDT funds on the Aleph chain");
     client.approve_usdt_azero(transfer_amount).await?;
 
     info!("Request the transfer of USDT to the Ethereum chain");
@@ -106,6 +82,12 @@ pub async fn usdt_to_usdt() -> Result<()> {
     .await
 }
 
+/// One-way `Aleph Zero` -> `Ethereum` transfer through `most`.
+/// Wraps AZERO into wAZERO, and approves the `most` contract to use the wAZERO funds.
+/// Burns the required funds in the wAZERO contract on Aleph Zero.
+/// Transfers `transfer_amount` of burned wAZERO over the bridge.
+/// Waits for the transfer to complete.
+/// It relies on all the relevant contracts being deployed on both ends and the (wAZERO_ETH:wAZERO_AZERO) pair having been added to `most`.
 #[tokio::test]
 pub async fn wazero_to_wazero() -> Result<()> {
     let config = setup_test();
@@ -114,12 +96,12 @@ pub async fn wazero_to_wazero() -> Result<()> {
     let client = Client::new(test_context);
     let initial_balance = client.balance().await?;
 
-    info!("{:?}", initial_balance);
+    info!("Initial balance: {:?}", initial_balance);
 
-    info!("wrap azero");
+    info!("Wrap native AZERO into wAZERO");
     client.wrap_wazero(transfer_amount).await?;
 
-    info!("Approve the `most` contract to use the wAZERO funds on the Azero chain");
+    info!("Approve the `most` contract to use the wAZERO funds on the Aleph chain");
     client.approve_wazero_azero(transfer_amount).await?;
 
     info!("Request the transfer of wAZERO to the Ethereum chain");
@@ -144,6 +126,49 @@ pub async fn wazero_to_wazero() -> Result<()> {
     .await
 }
 
+/// One-way `Aleph Zero` -> `Ethereum` transfer through `most`.
+/// Requires a prior transaction in the other direction to have completed.
+/// This is easily done by running the test for the other direction first.
+/// Approves the `most` contract to use the wETH funds.
+/// Burns the required funds in the wETH contract on Aleph Zero.
+/// Transfers `transfer_amount` of burned wETH over the bridge, the Ethereum account receives native ETH.
+/// Waits for the transfer to complete.
+/// It relies on all the relevant contracts being deployed on both ends and the (wETH_ETH:wETH_AZERO) pair having been added to `most`.
+#[tokio::test]
+pub async fn weth_to_eth() -> Result<()> {
+    let config = setup_test();
+    let test_context = config.create_test_context().await?;
+    let transfer_amount = utils::parse_ether(config.test_args.transfer_amount.clone())?.as_u128();
+    let client = Client::new(test_context);
+    let initial_balance = client.balance().await?;
+
+    info!("Initial balance: {:?}", initial_balance);
+
+    info!("Approve the `most` contract to use the wETH funds on the Aleph chain");
+    client.approve_weth_azero(transfer_amount).await?;
+
+    info!("Request the transfer of wETH to the Ethereum chain");
+    client.request_eth_transfer_azero(transfer_amount).await?;
+
+    info!("Wait for balance change");
+    let target_balance = initial_balance.bridge_eth_azero_to_eth(transfer_amount)?;
+    info!("Target balance: {:?}", target_balance);
+
+    let get_current_balance = || async { client.balance().await };
+    wait_for_balance_change(
+        get_current_balance,
+        target_balance,
+        Some(0.into()),
+        None,
+        config.test_args.wait_max_minutes,
+    )
+    .await
+}
+
+/// One-way `Aleph Zero` -> `Ethereum` transfer through `most`.
+/// Transfers `transfer_amount` of native AZERO over the bridge.
+/// Waits for the transfer to complete.
+/// It relies on all the relevant contracts being deployed on both ends and the (wAZERO_ETH:wAZERO_AZERO) pair having been added to `most`.
 #[tokio::test]
 pub async fn azero_to_wazero() -> Result<()> {
     let config = setup_test();
@@ -152,7 +177,7 @@ pub async fn azero_to_wazero() -> Result<()> {
     let client = Client::new(test_context);
     let initial_balance = client.balance().await?;
 
-    info!("{:?}", initial_balance);
+    info!("Initial balance: {:?}", initial_balance);
 
     info!("Request the transfer of AZERO to the Ethereum chain");
     client.request_azero_transfer_azero(transfer_amount).await?;
